Fix value parsing and routing for dynamic impulses.

+++ Remora.Neos.Headless.API.Abstractions/INeosWorldController.cs
@@ -200,7 +200,7 @@ public interface INeosWorldController
     (
         string worldId,
         string tag,
-        OneOf<float, int, string>? value = null,
+        OneOf<int, float, string>? value = null,
         CancellationToken ct = default
     );
 }
